+++ src/c/lua.c
@@ -146,14 +146,14 @@ NODISCARD static error push_files(lua_State *const L, struct wstr *const pattern
         (fd.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) == FILE_ATTRIBUTE_DIRECTORY) {
       continue;
     }
-    ptrdiff_t extpos = 0;
+    size_t extpos = 0;
     struct wstr ws = wstr_unmanaged(fd.cFileName);
     err = extract_file_extension(&ws, &extpos);
     if (efailed(err)) {
       err = ethru(err);
       goto cleanup;
     }
-    ws.len = (size_t)extpos;
+    ws.len = extpos;
     ws.ptr[extpos] = L'\0';
     err = luafn_push_wstr(L, &ws);
     if (efailed(err)) {
